requirements.txt
openai-agents[litellm]